refactor: make definition of the execute method consistent for all steps
