+++ classes/local/execution/connector_engine_step.php
@@ -44,9 +44,9 @@ class connector_engine_step extends engine_step {
         switch ($this->proceed_status()) {
             case self::PROCEED_GO:
                 try {
-                    $result = $this->steptype->execute($this);
+                    $result = $this->steptype->execute(new \stdClass);
                     $this->steptype->prepare_vars();
-                    if ($result === true) {
+                    if ($result !== false) {
                         $this->set_status(engine::STATUS_FINISHED);
                     } else {
                         $this->set_status(engine::STATUS_CANCELLED);
