+++ classes/local/step/connector_debug_file_display.php
@@ -35,7 +35,7 @@ class connector_debug_file_display extends connector_step {
      *
      * @return bool Returns true if successful, false otherwise.
      */
-    public function execute(): bool {
+    public function execute($input = null) {
         $config = $this->enginestep->stepdef->config;
 
         $streamname = $this->enginestep->engine->resolve_path($config->streamname);
