+++ classes/local/step/connector_sns_notify.php
@@ -71,7 +71,7 @@ class connector_sns_notify extends connector_step {
      *
      * @return bool Returns true if successful, false otherwise.
      */
-    public function execute(): bool {
+    public function execute($input = null) {
         global $CFG;
 
         try {
