+++ classes/local/step/connector_wait.php
@@ -57,7 +57,7 @@ class connector_wait extends connector_step {
      *
      * @return bool Returns true if successful, false otherwise.
      */
-    public function execute(): bool {
+    public function execute($input = null) {
         $timesec = $this->stepdef->config->timesec;
         if (!is_int($timesec) && !(is_string($timesec) && ctype_digit($timesec))) {
             throw new \moodle_exception('connector_wait:not_integer', 'tool_dataflows', '', $timesec);
