+++ classes/local/step/trigger_step.php
@@ -45,7 +45,7 @@ abstract class trigger_step extends connector_step {
      *
      * @return true for now, since this will be configured differently per step
      */
-    public function execute(): bool {
+    public function execute($input = null) {
         return true;
     }
